chore: modify contact button text
the contact backend is currently not avaliable as i have not one goddamn
idea how to make a personal mailbox bot and send me all those messages.
I also have not one single clue in mind how to deploy my node.js backend
on to my server. I also have not one single clue on how to use
serverless as an alternative of deploying the code to server. So save
me, anyone!